# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/wines/{id}":{"get":{"tags":["wine-controller"],"operationId":"getWineById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WineResponse"}}}}}},"put":{"tags":["wine-controller"],"operationId":"updateWine","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WineUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WineResponse"}}}}}},"delete":{"tags":["wine-controller"],"operationId":"deleteWine","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/events/{id}":{"get":{"tags":["event-controller"],"operationId":"getEventById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}},"put":{"tags":["event-controller"],"operationId":"updateEvent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}},"delete":{"tags":["event-controller"],"operationId":"deleteEvent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/courses/{id}":{"get":{"tags":["course-controller"],"operationId":"getCourseById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}}}},"put":{"tags":["course-controller"],"operationId":"updateCourse","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}}}},"delete":{"tags":["course-controller"],"operationId":"deleteCourse","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/wines":{"get":{"tags":["wine-controller"],"operationId":"listWines","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelWineResponse"}}}}}},"post":{"tags":["wine-controller"],"operationId":"createWine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WineUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WineResponse"}}}}}}},"/api/events":{"get":{"tags":["event-controller"],"operationId":"listEvents","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelEventResponse"}}}}}},"post":{"tags":["event-controller"],"operationId":"createEvent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/events/{eventId}/courses":{"get":{"tags":["course-controller"],"operationId":"listCoursesByEvent","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelCourseResponse"}}}}}},"post":{"tags":["course-controller"],"operationId":"createCourse","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}}}}},"/api/events/{eventId}/wines":{"get":{"tags":["wine-controller"],"operationId":"listWinesByEvent","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelWineResponse"}}}}}}},"/api/courses":{"get":{"tags":["course-controller"],"operationId":"listCourses","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelCourseResponse"}}}}}}}},"components":{"schemas":{"WineUpsertRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"type":{"type":"string","enum":["WHITE","RED","ROSE","SPARKLING","DESSERT","FORTIFIED","UNKNOWN"]},"grape":{"type":"string","minLength":1},"country":{"type":"string","minLength":1},"region":{"type":"string"},"year":{"type":"string"}},"required":["country","grape","name","type"]},"CourseResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"event":{"$ref":"#/components/schemas/EventResponse"},"courseNo":{"type":"integer","format":"int32"},"cook":{"type":"string"},"dish":{"$ref":"#/components/schemas/DishResponse"}}},"DishResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"mainIngredient":{"type":"string"}}},"EventResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"host":{"type":"string"},"location":{"type":"string"}}},"WineResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string","enum":["WHITE","RED","ROSE","SPARKLING","DESSERT","FORTIFIED","UNKNOWN"]},"grape":{"type":"string"},"country":{"type":"string"},"region":{"type":"string"},"year":{"type":"string"},"courses":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponse"}}}},"EventUpsertRequest":{"type":"object","properties":{"date":{"type":"string","format":"date"},"host":{"type":"string","minLength":1},"location":{"type":"string"}},"required":["date","host"]},"CourseUpsertRequest":{"type":"object","properties":{"courseNo":{"type":"integer","format":"int32"},"cook":{"type":"string","minLength":1},"dish":{"$ref":"#/components/schemas/DishUpsertRequest"}},"required":["cook","courseNo"]},"DishUpsertRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"mainIngredient":{"type":"string"}},"required":["name"]},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModelWineResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/WineResponse"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"PagedModelEventResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/EventResponse"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"PagedModelCourseResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponse"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}}}}}